chore: upgrade multi-nicd ginkgo v2
