+++ daemon/dockerfiles/Dockerfile.multi-nicd-test
@@ -1,5 +1,9 @@
 FROM ghcr.io/foundation-model-stack/multi-nic-cni-kbuilder:v1.2.9
 
+RUN  apt-get update \
+  && apt-get install -y jq \
+  && rm -rf /var/lib/apt/lists/*
+
 RUN mkdir -p /usr/local/app
 RUN mkdir -p /host/opt/cni/bin
 
@@ -18,12 +22,9 @@ RUN cd cni && go mod tidy && go build -o /usr/local/app/aws-ipvlan ./plugins/mai
 RUN cd cni && make ginkgo-set
 # build Daemon
 COPY daemon daemon
+COPY hack daemon/hack
 RUN mkdir -p /usr/local/kubebuilder
 RUN mkdir -p /usr/local/build/daemon/src/test-bin
-RUN curl -sSLo setup-envtest.sh https://raw.githubusercontent.com/kubernetes-sigs/controller-runtime/v0.7.2/hack/setup-envtest.sh
-RUN cp setup-envtest.sh /usr/local/build/daemon/src/test-bin/setup-envtest.sh
-RUN cd daemon/src && make test-env && cp -r /usr/local/build/daemon/src/test-bin/bin /usr/local/kubebuilder/bin
-RUN mkdir -p daemon/src/test-bin
 RUN cd daemon/src && go mod tidy
 RUN cd daemon/src && go build -o /usr/local/app/daemon
 RUN cp /usr/local/app/multi-nic /usr/local/bin \
